+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # @melt-ui/svelte
 
+## 0.24.0
+
+### Minor Changes
+
+- ef32b32a: [New Builder] Toast
+
+### Patch Changes
+
+- f25ed135: [Dialog]: Dialog no longer closes on right-click
+
 ## 0.23.2
 
 ### Patch Changes